+++ main.py
@@ -14,11 +14,9 @@ def create_app():
     from routes.chart_route import bp as chart_bp
     from routes.dashboard_route import bp as dashboard_bp
     from routes.trend_route import bp as trend_bp
-    from routes.sentiment_route import bp as sentiment_bp
     app.register_blueprint(chart_bp)
     app.register_blueprint(dashboard_bp)
     app.register_blueprint(trend_bp)
-    app.register_blueprint(sentiment_bp)
 
     @app.route('/')
     def home_page():
