+++ static/screen/dashboard.css
@@ -6,8 +6,14 @@
     min-width: 0;
     word-wrap: break-word;
     background-color: #fff;
-    border-radius: 20px;
+    border: none;
+    box-shadow: 0 .5rem 1rem rgba(0,0,0,.15)!important;
 }
+
+.h2 {
+    font-weight: bolder;
+    color: darkgrey;
+    }
 
 .card-title,
 .card-content{
