+++ templates/dashboard_base.html
@@ -129,7 +129,7 @@
 
                 <li class="nav-item">
                       <div class="sentiment">
-                          <a href="/sentiment"><button type="button" class="btn btn-outline-secondary btn-sm">감성분석</button></a>
+                          <a href="/dashboard/sentiment"><button type="button" class="btn btn-outline-secondary btn-sm">감성분석</button></a>
                       </div>
                 </li>
